+++ client/src/components/index.ts
@@ -1,0 +1,15 @@
+import AlertMessage from "./AlertMessage";
+import NavigationBar from "./NavigationBar";
+import PrimaryButton from "./PrimaryButton";
+import SearchBar from "./SearchBar";
+import SingleProduct from "./SingleProduct";
+import ThinComponent from "./ThinComponent";
+
+export {
+	AlertMessage,
+	NavigationBar,
+	PrimaryButton,
+	SearchBar,
+	SingleProduct,
+	ThinComponent,
+};
